Renamed to IR builder for better PR readability
