+++ .gitignore
@@ -1,4 +1,4 @@
-*build*
+build*
 spconfig.*
 *~
 .DS_Store
